requirements.txt
funcsigs==0.4
ipfs-api==0.2.2
mock==1.0.1
pbr==1.8.1
pycrypto==2.6.1
requests==2.9.1
six==1.10.0
wheel==0.24.0

